feat: move combinedTickers into socketStore
